Refactor: task 1
