+++ src/hw1/task1/Task1.java
@@ -1,4 +1,4 @@
-package hw1;
+package hw1.task1;
 
 import java.util.Arrays;
 import java.util.List;
